refactor: move parseDataRate function to utils file
